pyproject.toml
[project]
name = "book-rag"
version = "0.2.0"
description = "RAG chatbot for books - upload any book and chat with it"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.115.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.0",
    "ollama>=0.6.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.53.0",
    "uvicorn>=0.34.0",
]
